+++ packages/scenes/src/components/EmbeddedScene.tsx
@@ -5,7 +5,6 @@ import React from 'react';
 
 import { SceneObjectBase } from '../core/SceneObjectBase';
 import { SceneComponentProps, SceneObjectState, SceneObject } from '../core/types';
-import { getUrlSyncManager } from '../services/UrlSyncManager';
 import { setWindowGrafanaSceneContext } from '../utils/compatibility/setWindowGrafanaSceneContext';
 
 export interface EmbeddedSceneState extends SceneObjectState {
@@ -31,19 +30,8 @@ export class EmbeddedScene extends SceneObjectBase<EmbeddedSceneState> {
       const unsetGlobalScene = setWindowGrafanaSceneContext(this);
       return () => {
         unsetGlobalScene();
-        getUrlSyncManager().cleanUp(this);
       };
     });
-  }
-
-  /**
-   * initUrlSync should be called before the scene is rendered to ensure that objects are in sync
-   * before they get activated. This saves some unnecessary re-renders and makes sure variables
-   * queries are issued as needed. If your using SceneAppPage you will not need to call this as
-   * url sync is handled on the SceneAppPage level not this level.
-   */
-  public initUrlSync() {
-    getUrlSyncManager().initSync(this);
   }
 }