+++ packages/scenes/src/variables/adhoc/AdHocFilterRenderer.tsx
@@ -17,7 +17,7 @@ function selectableValueToVariableValueOption(value: SelectableValue): VariableV
   return {
     label: value.label ?? String(value.value),
     value: value.value,
-  }
+  };
 }
 
 function keyLabelToOption(key: string, label?: string): SelectableValue | null {
@@ -77,7 +77,7 @@ export function AdHocFilterRenderer({ filter, model }: Props) {
       inputValue={valueInputValue}
       onInputChange={onValueInputChange}
       onChange={(v) => {
-        model._updateFilter(filter, 'value', v)
+        model._updateFilter(filter, 'value', v);
 
         if (valueHasCustomValue !== v.__isNew__) {
           setValueHasCustomValue(v.__isNew__);
@@ -152,7 +152,17 @@ export function AdHocFilterRenderer({ filter, model }: Props) {
 
       return (
         <Field label={label} data-testid={`AdHocFilter-${filter.key}`} className={styles.field}>
-          {valueSelect}
+          <div className={styles.wrapper}>
+            <Select
+              className={styles.operator}
+              value={filter.operator}
+              disabled={model.state.readOnly}
+              options={model._getOperators()}
+              width="auto"
+              onChange={(v) => model._updateFilter(filter, 'operator', v)}
+            />
+            {valueSelect}
+          </div>
         </Field>
       );
     } else {